# openapi.json
{
    "openapi": "3.0.0",
    "info": {
      "title": "Combined API Specification",
      "version": "1.0.0"
    },
    "paths": {
...
      "/goal-service/goals": {
        "get": {
          "summary": "Get Goals",
          "operationId": "get_goals_goals_get",
          "responses": {
            "200": {
              "description": "Successful Response",
              "content": {
                "application/json": {
                  "schema": {
                    "items": {
                      "$ref": "#/components/schemas/Goal"
                    },
                    "type": "array",
                    "title": "Response Get Goals Goals Get"
                  }
                }
              }
            }
          },
          "security": [
            {
              "OAuth2PasswordBearer": []
            }
          ]
        },
        "post": {
          "summary": "Create Goal",
          "operationId": "create_goal_goals_post",
          "requestBody": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GoalCreate"
                }
              }
            },
            "required": true
          },
          "responses": {
            "200": {
              "description": "Successful Response",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/Goal"
                  }
                }
              }
            },
            "422": {
              "description": "Validation Error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/HTTPValidationError"
                  }
                }
              }
            }
          },
          "security": [
            {
              "OAuth2PasswordBearer": []
            }
          ]
        }
      },
      "/goal-service/goals/{goal_id}/tasks": {
        "post": {
          "summary": "Create Task For Goal",
          "operationId": "create_task_for_goal_goals__goal_id__tasks_post",
          "security": [
            {
              "OAuth2PasswordBearer": []
            }
          ],
          "parameters": [
            {
              "name": "goal_id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "integer",
                "title": "Goal Id"
              }
            }
          ],
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskCreate"
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Successful Response",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/Task"
                  }
                }
              }
            },
            "422": {
              "description": "Validation Error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/HTTPValidationError"
                  }
                }
              }
            }
          }
        },
        "get": {
          "summary": "Get Tasks For Goal",
          "operationId": "get_tasks_for_goal_goals__goal_id__tasks_get",
          "security": [
            {
              "OAuth2PasswordBearer": []
            }
          ],
          "parameters": [
            {
              "name": "goal_id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "integer",
                "title": "Goal Id"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Successful Response",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/Task"
                    },
                    "title": "Response Get Tasks For Goal Goals  Goal Id  Tasks Get"
                  }
                }
              }
            },
            "422": {
              "description": "Validation Error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/HTTPValidationError"
                  }
                }
              }
            }
          }
        }
      },
      "/goal-service/tasks/{task_id}": {
        "put": {
          "summary": "Update Task",
          "operationId": "update_task_tasks__task_id__put",
          "security": [
            {
              "OAuth2PasswordBearer": []
            }
          ],
          "parameters": [
            {
              "name": "task_id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "integer",
                "title": "Task Id"
              }
            }
          ],
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Task"
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Successful Response",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/Task"
                  }
                }
              }
            },
            "422": {
              "description": "Validation Error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/HTTPValidationError"
                  }
                }
              }
            }
          }
        },
        "delete": {
          "summary": "Delete Task",
          "operationId": "delete_task_tasks__task_id__delete",
          "security": [
            {
              "OAuth2PasswordBearer": []
            }
          ],
          "parameters": [
            {
              "name": "task_id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "integer",
                "title": "Task Id"
              }
            }
          ],
          "responses": {
            "204": {
              "description": "Successful Response"
            },
            "422": {
              "description": "Validation Error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/HTTPValidationError"
                  }
                }
              }
            }
          }
        }
      }
    },
    "components": {
      "schemas": {
...
        "HTTPValidationError": {
          "properties": {
            "detail": {
              "items": {
                "$ref": "#/components/schemas/ValidationError"
              },
              "type": "array",
              "title": "Detail"
            }
          },
          "type": "object",
          "title": "HTTPValidationError"
        },
...
        "ValidationError": {
          "properties": {
            "loc": {
              "items": {
                "anyOf": [
                  {
                    "type": "string"
                  },
                  {
                    "type": "integer"
                  }
                ]
              },
              "type": "array",
              "title": "Location"
            },
            "msg": {
              "type": "string",
              "title": "Message"
            },
            "type": {
              "type": "string",
              "title": "Error Type"
            }
          },
          "type": "object",
          "required": [
            "loc",
            "msg",
            "type"
          ],
          "title": "ValidationError"
        },
        "Goal": {
          "properties": {
            "id": {
              "type": "integer",
              "title": "Id"
            },
            "title": {
              "type": "string",
              "title": "Title"
            },
            "description": {
              "type": "string",
              "title": "Description"
            }
          },
          "type": "object",
          "required": [
            "id",
            "title",
            "description"
          ],
          "title": "Goal"
        },
        "GoalCreate": {
          "properties": {
            "id": {
              "type": "integer",
              "title": "Id"
            },
            "title": {
              "type": "string",
              "title": "Title"
            },
            "description": {
              "type": "string",
              "title": "Description"
            }
          },
          "type": "object",
          "required": [
            "id",
            "title",
            "description"
          ],
          "title": "GoalCreate"
        },
        "Task": {
          "properties": {
            "id": {
              "type": "integer",
              "title": "Id"
            },
            "title": {
              "type": "string",
              "title": "Title"
            },
            "description": {
              "type": "string",
              "title": "Description"
            },
            "status": {
              "type": "string",
              "title": "Status"
            },
            "goal_id": {
              "type": "integer",
              "title": "Goal Id"
            },
            "assignee": {
              "type": "string",
              "title": "Assignee"
            }
          },
          "type": "object",
          "required": [
            "id",
            "title",
            "description",
            "status",
            "goal_id",
            "assignee"
          ],
          "title": "Task"
        },
        "TaskCreate": {
          "properties": {
            "id": {
              "type": "integer",
              "title": "Id"
            },
            "title": {
              "type": "string",
              "title": "Title"
            },
            "description": {
              "type": "string",
              "title": "Description"
            },
            "status": {
              "type": "string",
              "title": "Status"
            },
            "assignee": {
              "type": "string",
              "title": "Assignee"
            }
          },
          "type": "object",
          "required": [
            "id",
            "title",
            "description",
            "status",
            "assignee"
          ],
          "title": "TaskCreate"
        }
      },
      "securitySchemes": {
        "OAuth2PasswordBearer": {
          "type": "oauth2",
          "flows": {
            "password": {
              "scopes": {},
              "tokenUrl": "http://localhost:8001/token"
            }
          }
        }
      }
    },
    "security": []
  }